Remove wireSaveData cache to resolve sync issues
The wireSaveData cache did not sync properly with the CircuitStore. This resulted in wire manipulations of one circuit being present in others, along with deleted wires still being present in the save data. There is seemingly no reason to use the intermediate cache here rather than modifying the CircuitStore directly.